backend/shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
description = "poxa calc service backend shared"
requires-python = "==3.11.9"
dependencies = [
    "pydantic>=2.0.0",
    "pyhumps>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["models", "schemas"]


[tool.uv.sources]
sst = { git = "https://github.com/sst/sst.git", subdirectory = "sdk/python", branch = "dev" }
